Remove "Followed Podcasts" carousel.
