+++ app/src/main/java/com/august/jetcaster/util/DateTime.kt
@@ -1,0 +1,25 @@
+package com.august.jetcaster.util
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.res.stringResource
+import com.august.jetcaster.R
+import java.time.Duration
+import java.time.LocalDateTime
+import java.time.OffsetDateTime
+
+@Composable
+private fun lastUpdated(updated: OffsetDateTime): String {
+    val duration = Duration.between(updated.toLocalDateTime(), LocalDateTime.now())
+    val days = duration.toDays().toInt()
+
+    return when {
+        days > 28 -> stringResource(R.string.updated_longer)
+        days >= 7 -> {
+            val weeks = days / 7
+            quantityStringResource(R.plurals.updated_weeks_ago, weeks, weeks)
+        }
+
+        days > 0 -> quantityStringResource(R.plurals.updated_days_ago, days, days)
+        else -> stringResource(R.string.updated_today)
+    }
+}
